pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "google-api-python-client>=2.183.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.15.0",
    "openai-agents>=0.3.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
